// packages/hooks/src/useRequest/types.ts
export type UseRequestService<TData, TParams extends unknown[]> = (
  ...args: TParams
) => Promise<TData>;

export interface UseRequestFetchState<TData, TParams extends unknown[]> {
  loading: boolean;
  params?: TParams;
  data?: TData;
  error?: Error | unknown;
}

export type UseRequestOptions<TData, TParams extends unknown[], TPlugin> = {
  [K in keyof UseRequestBasicOptions<TData, TParams>]: UseRequestBasicOptions<TData, TParams>[K];
} & {
  [K in keyof TPlugin]: TPlugin[K];
};

export interface UseRequestPlugin<TData, TParams extends unknown[] = unknown[], TPlugin = any> {
  (
    fetchInstance: UseRequestFetch<TData, TParams>,
    options: UseRequestOptions<TData, TParams, TPlugin>,
  ): UseRequestPluginReturn<TData, TParams>;
  onInit?: (
    options: UseRequestOptions<TData, TParams, TPlugin>,
  ) => Partial<UseRequestFetchState<TData, TParams>>;
}

export interface UseRequestPluginReturn<TData, TParams extends unknown[]> {
  name?: string;
  onBefore?: (params: TParams) =>
    | ({
        stopNow?: boolean;
        returnNow?: boolean;
      } & Partial<UseRequestFetchState<TData, TParams>>)
    | void;

  onRequest?: (
    service: UseRequestService<TData, TParams>,
    params: TParams,
  ) => {
    servicePromise?: Promise<TData>;
  };

  onSuccess?: (data: TData, params: TParams) => void;
  onError?: (e: Error, params: TParams) => void;
  onFinally?: (params: TParams, data?: TData, e?: Error) => void;
  onCancel?: () => void;
  onMutate?: (data: TData) => void;
}